modify manager

+++ scripts/manager.py
@@ -2,7 +2,7 @@
 import sys
 sys.path.append('.')
 
-from flask.ext.script import Manager, Shell
+from flask.ext.script import Manager, Shell, prompt_bool
 from flask.ext.migrate import Migrate, MigrateCommand
 '''
     add by daisheng
@@ -12,7 +12,8 @@ from sqlite3 import dbapi2 as sqlite3
 from tecstack import app, db
 from tecstack import auth
 '''
-    add by Leann Mak 2015/7/5
+    add by leannmak
+    2015/7/5
 '''
 from tecstack import vminfo
 '''
@@ -20,13 +21,20 @@ from tecstack import vminfo
 '''
 
 migrate = Migrate(app, db)
-
-manager = Manager(app)
+'''
+    add usage by leannmak
+    2015/7/13
+'''
+manager = Manager(app, usage="Perform database operations")
+'''
+    end
+'''
 manager.add_command('db', MigrateCommand)
 
 
 @manager.command
 def initdb():
+    "initialize database tables"
     import os
     if not os.path.exists(app.config['DB_FOLDER']):
         os.mkdir(app.config['DB_FOLDER'])
@@ -40,13 +48,15 @@ def initdb():
 
 @manager.command
 def importdata():
+    "Import data into database tables"
     rv = sqlite3.connect(app.config['DB_FILEPATH'])
     rv.row_factory = sqlite3.Row
     with app.open_resource(app.config['DB_SOURCEFILEPATH'], mode='r') as f:
         rv.cursor().executescript(f.read())
     rv.commit()
     '''
-        add by Leann Mak 2015/7/5
+        add by leannmak
+        2015/7/5
     '''
     rv.close()
     '''
@@ -58,8 +68,32 @@ def importdata():
 
 @manager.command
 def dropdb():
-    db.drop_all()
-    print 'Database droped.'
+    '''
+        add prompt by leannmak
+        2015/7/13
+    '''
+    "Drops database tables"
+    if prompt_bool("Are you sure you want to lose all your data"):
+        db.drop_all()
+        print 'Database droped.'
+
+
+'''
+    add by leannmak
+    2015/7/13
+'''
+
+
+@manager.command
+def recreatedb():
+    "Recreates database tables \
+    (same as issuing 'drop', 'create' and then 'import')"
+    dropdb()
+    initdb()
+    importdata()
+'''
+    end
+'''
 
 
 def _make_context():
